backend/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Google Cloud & Vertex AI
google-cloud-aiplatform>=1.42.0
google-cloud-storage>=2.14.0
google-generativeai>=0.3.0
google-genai>=0.6.0
google-cloud-discoveryengine>=0.13.0  # Vertex AI Search

# LangChain & LangGraph for multi-agent document validation
langgraph>=0.2.0
langchain-google-vertexai>=2.0.0
langchain-core>=0.3.0

# Database
sqlalchemy>=2.0.25
alembic>=1.13.0

# Utilities
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dateutil>=2.8.2

# PDF and document processing
PyPDF2>=3.0.0
pillow>=10.3.0
pytesseract>=0.3.10

# HTTP client
httpx>=0.26.0

# Cloud Run / Production
gunicorn>=21.0.0
